Moving data of saved rules to a separate component from the template view
